refactor: meeting, member package 위치 변경
