+++ src/main/java/com/mobble/mobbleserver/domain/meeting/repository/MeetingRepository.java
@@ -1,6 +1,6 @@
-package com.mobble.mobbleserver.meeting.repository;
+package com.mobble.mobbleserver.domain.meeting.repository;
 
-import com.mobble.mobbleserver.meeting.entity.Meeting;
+import com.mobble.mobbleserver.domain.meeting.entity.Meeting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MeetingRepository extends JpaRepository<Meeting, Long> {
