+++ src/main/java/com/mobble/mobbleserver/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.mobble.mobbleserver.member.repository;
+package com.mobble.mobbleserver.domain.member.repository;
 
-import com.mobble.mobbleserver.member.entity.Member;
+import com.mobble.mobbleserver.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
